requirements.txt
tensorflow-gpu
sentencepiece
